refactored project and docs
