+++ docs/gh-security-config-demo.gif
