+++ example-organizations.csv
@@ -1,5 +1,8 @@
 acme-corp
 tech-solutions
 data-analytics
+first-org1
+first-org10
+first-org11
 security-team
 infrastructure
